# CityPharmacyChain.Api.Client/Api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CityPharmacyChain.Api.Host",
    "version": "1.0"
  },
  "paths": {
...
    "/PriceListEntry": {
      "get": {
        "tags": [
          "PriceListEntry"
        ],
        "summary": "GET запрос по получению всех объектов класса запись в прайс-листе из базы данных",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PriceListEntryFullDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PriceListEntryFullDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PriceListEntryFullDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "PriceListEntry"
        ],
        "summary": "POST запрос по добавлению объекта класса запись в прайс-листе в базу данных",
        "requestBody": {
          "description": "Объект класса запись в прайс-листе в формате JSON",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PriceListEntryDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PriceListEntryDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PriceListEntryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PriceListEntryFullDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceListEntryFullDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceListEntryFullDto"
                }
              }
            }
          }
        }
      }
    },
    "/PriceListEntry/{id}": {
      "get": {
        "tags": [
          "PriceListEntry"
        ],
        "summary": "GET запрос по получению объекта класса запись в прайс-листе из базы данных по его идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор записи в прайс-листе",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PriceListEntryDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceListEntryDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceListEntryDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "PriceListEntry"
        ],
        "summary": "PUT запрос по модификации объекта класса запись в прайс-листе из базы данных по его идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор записи в прайс-листе",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Объект класса запись в прайс-листе в формате JSON",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PriceListEntryDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PriceListEntryDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PriceListEntryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PriceListEntryFullDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceListEntryFullDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceListEntryFullDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "PriceListEntry"
        ],
        "summary": "DELETE запрос по удалению существующего объекта класса запись в прайс-листе из базы данных по его идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор записи в прайс-листе",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PriceListEntryDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "description": "Идентификатор препарата",
            "format": "int32"
          },
          "pharmacyId": {
            "type": "integer",
            "description": "Идентификатор аптеки",
            "format": "int32"
          },
          "soldCount": {
            "maximum": 1000,
            "minimum": 0,
            "type": "integer",
            "description": "Количество проданных препаратов",
            "format": "int32",
            "nullable": true
          },
          "manufacturer": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "description": "Производитель препарата",
            "nullable": true
          },
          "paymentType": {
            "type": "string",
            "description": "Способ оплаты (наличные/картой)",
            "nullable": true
          },
          "saleDate": {
            "type": "string",
            "description": "Дата продажи препарата",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Класс DTO для сущности класса запись в прайс-листе"
      },
      "PriceListEntryFullDto": {
        "type": "object",
        "properties": {
          "priceListEntryId": {
            "type": "integer",
            "description": "Идентификатор записи в прайс-листе",
            "format": "int32"
          },
          "productId": {
            "type": "integer",
            "description": "Идентификатор препарата",
            "format": "int32"
          },
          "pharmacyId": {
            "type": "integer",
            "description": "Идентификатор аптеки",
            "format": "int32"
          },
          "soldCount": {
            "maximum": 1000,
            "minimum": 0,
            "type": "integer",
            "description": "Количество проданных препаратов",
            "format": "int32",
            "nullable": true
          },
          "manufacturer": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "description": "Производитель препарата",
            "nullable": true
          },
          "paymentType": {
            "type": "string",
            "description": "Способ оплаты (наличные/картой)",
            "nullable": true
          },
          "saleDate": {
            "type": "string",
            "description": "Дата продажи препарата",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Класс DTO для сущности класса запись в прайс-листе, содержащий её идентификатор"
      },
...
    }
  }
}